tung tung

+++ LudenWebAPI/Entities/Models/BillItem.cs
@@ -10,11 +10,11 @@ namespace Entities.Models
         [Key]
         public ulong Id { get; set; }
 
-        public int BillId { get; set; }
+        public ulong BillId { get; set; }
         [JsonIgnore]
         public Bill Bill { get; set; }
 
-        public int ProductId { get; set; }
+        public ulong ProductId { get; set; }
         public Product Product { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
